chore: Add README.md for project documentation
A README.md file has been added to the project to provide users and developers a comprehensive overview. It includes the purpose of the application, a preview screenshot, an enumerated list of features, used technologies, instructions for cloning and running the application, contributing guide, license information, and acknowledgements. This document will assist in onboarding new developers and helping users understand the application functionality and setup.